1) Added an outer game 'loop' by disabling/enabling the relevant bits of UI depending on whether a game is underway ('playing') or whether the program is waiting for a new game to start ('waiting'). 2) Made the minimal grid 9×9. 3) Added medium and hard modes.
I am not sure if this is the right way to do it, because it seems rather clunky and requires a lot of admin. Some hasty googles on how to write an event driven game in Python or in general does seem to point in this direction though. Maybe that within Tk things should be done differently.

Also the game seems to have gotten a touch slower, which is not what should happen when you are programming a type of game that should be able to run on computers a million times slower. Especially destroying a large board with 480 fields (button widgets) seems to be a sluggish process.

Another disadvantage of this approach is that I end up juggling a ton of global variables. I am sure something can be done about this. A crude idea would be, if I only need to read certain variables, I can do something ugly like 'def getMyvalue(): global myValue; return myValue'). Ultimately I hope though that as interfaces mature, I can offload functionality to other functions.