Tried to make the game look, if not prettier then at least clearer, by making the win and loss states clearer: 1) using bolded, coloured win and loss messages, 2) Render all flags on a win, 3) make the status bar less jerky between states by locking in a static height.
I need to come up with a better solution for clearing a field too: currently, when clearing a field I make clear that the field has been cleared by deleting the border: Tkinter sees that as a hint to make the button smaller, which can have a jerky effect at the end of the game on the whole window as that too is being resized.

Other things in the look and feel department also need to be dealt with, but I will tackle those once they start to annoy me sufficiently.