1) Added a game clock. 2) Reduced jerkiness by drawing both button states in the same size.
With regards to the redrawing, the default for Tkinter is to only use as much space as it needs. I currently draw a cleared field by removing the border from the button. This decreases the width and height of the button by 4 pixels each, so at the same time I need to increase the width and height of the now borderless button both by 4 pixels.

Speaking of drawing things, I have done some speed tests on drawing and clearing a board: drawing my own buttons using Canvas() makes both drawing and clearing the board about an order of magnitude faster. That may be something to look into in the future, especially since clearing a large board of Tkinter's buttons takes over half a second, which is a noticeable delay (especially with all the admin going on at the same time).