1) Removed some unused variables, renamed others. 2) Call openNeighbours() with the testGrid variable instead of a copy.
Turns out it was a shallow copy anyway, I don't yet have a deep understanding of Python. The variable testGrid gets cleared every game, it makes no sense to then make a copy. I am not even sure if keeping track of tested fields like this makes sense, the game tends to operate in the neighbourhood of fields that haven't been cleared yet anyway, plus we can always query uiTree[y][x].visible to find out if a field has been tested. I have tried a few runs of the game with and without storing the tested status and it does not seem to make much difference.