I wrote I would only change certain things once I got sufficiently annoyed and I am keeping my promise. Although redrawing the board for a new game was slow at about 1.5 seconds if coming from the largest board, it seemed to be fine for the medium size, until I felt it wasn't.
So I have rewritten all code that worked with Tkinter's button widgets to work with the canvas widget instead.

It is still a bit ugly and a bit broken in some places, but you can currently play a game without getting too upset and the rest will have to follow in future commits.

One thing I noticed while working on this is that the drawing board does not seem to get properly destroyed so potentially this is a memory leak (I do not know how Python collects its garbage). This is something that still needs to be looked at.

So the bad news is that I now have to do a lot of the management that Tkinter took care of when I still used proper buttons, the good news is I could kill off several of the tricks and work-arounds I had to use because the Mines fields do not necessarily act like buttons.