1) Fix bug: cleared fields are now shown properly at the end of a game. 2) Fix bug: player can now plant and remove flags.
These were two left-overs from porting the game from a Tkinter Button-based version to a Tkinter Canvas-based version (faster, but harder to manage).

Wrt the flags, that is just something I had overlooked, because I rarely plant flags when I play (a good todo for a test suite), so this still needed to get the conversion treatment.

The problem with the cleared fields was that some of them had a border and some of them did not. This could have been the result of an aesthetic decision, but it was not. The problem arose when a field was cleared twice.