1) Started to split up functions to get looser coupling. 2) Added comments to each function outlining what the function does. 3) Renamed some of the functions in the hope the new name better represents what the function does or returns. 4) Fixed a small bug where Field.getCoords() possibly returned the wrong values. 5) Removed a couple of unused variables. 6) Renamed the global 'level' to 'start_level' as it was confusing me; I seemed to think 'level' represented the current level, which it does not (it may do so as a local variable inside functions though).
I am mulling over prefixing functions that mainly operate on the GUI with 'ui_' in order to make them stand out more.

For a program as small as this I am not terribly interested in proper cohesion and coupling, however I feel untangling functions and giving them clearer interfaces makes it easier to test them.

For example, addHighscore() would first be involved in the domains of time, file I/O and general logic, but now that it calls extendScorelist(highscores, score, today) I can feed the latter (which contains most of the logic of the former) with lots of yummy mock values rather than having to hope that the main program is in a certain state at the time of testing.